S3_Vector/requirements.txt
openai
boto3
PyPDF2
streamlit
python-dotenv
